instagram-scraper/requirements.txt
requests==2.32.3
python-dotenv==1.0.0
Pillow>=10.0.0
